default.html: make homepage header not be a link
On the homepage, the large header is a link back to the root. This
results in confusing behaviour for those who use the middle mouse button
to scroll, because attempting to scroll will open a new tab with the
same page, and fail to scroll. This link is not very useful anyway since
it leads to the page we are currently on. So it makes sense to just
remove the link for the homepage.